r/cognitive_account: purging when the feature flag is enabled

+++ azurerm/internal/services/cognitive/cognitive_account_resource.go
@@ -408,14 +408,17 @@ func resourceCognitiveAccountDelete(d *pluginsdk.ResourceData, meta interface{})
 		return fmt.Errorf("waiting for deletion of %s: %+v", *id, err)
 	}
 
-	// TODO: add a feature flag for this
-	log.Printf("[DEBUG] Purging %s..", *id)
-	purgeFuture, err := deletedAccountsClient.Purge(ctx, *account.Location, id.ResourceGroup, id.Name)
-	if err != nil {
-		return fmt.Errorf("purging %s: %+v", *id, err)
-	}
-	if err := purgeFuture.WaitForCompletionRef(ctx, deletedAccountsClient.Client); err != nil {
-		return fmt.Errorf("waiting for purge of %s: %+v", *id, err)
+	if meta.(*clients.Client).Features.CognitiveAccount.PurgeSoftDeleteOnDestroy {
+		log.Printf("[DEBUG] Purging %s..", *id)
+		purgeFuture, err := deletedAccountsClient.Purge(ctx, *account.Location, id.ResourceGroup, id.Name)
+		if err != nil {
+			return fmt.Errorf("purging %s: %+v", *id, err)
+		}
+		if err := purgeFuture.WaitForCompletionRef(ctx, deletedAccountsClient.Client); err != nil {
+			return fmt.Errorf("waiting for purge of %s: %+v", *id, err)
+		}
+	} else {
+		log.Printf("[DEBUG] Skipping Purge of %s", *id)
 	}
 
 	return nil
